working but not performant
