+++ packages/ui-components/src/common/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as usePromise } from "./usePromise";
 export * from "./metric-data";
+export { default as useMetricMapFillColor } from "./useMetricMapFillColor";
